+++ atomic_physics/julia_tools.py
@@ -1,0 +1,18 @@
+import sys
+
+try:
+    from juliacall import Main as jl
+
+    def fmt():
+        jl.seval("using JuliaFormatter")
+        jl.seval('format(".")')
+
+    def fmt_test():
+        jl.seval("using JuliaFormatter")
+        ret = jl.seval('format(".")')
+
+        if not ret:
+            sys.exit(1)
+
+except ModuleNotFoundError:
+    pass